+++ packages/contracts/src/index.js
@@ -1,2 +1,3 @@
 export * from './token_contracts'
 export * from './chain_contracts'
+export * from './bridge_contracts'